+++ src/components/chip-set/chip-set.tsx
@@ -638,6 +638,7 @@ export class ChipSet {
                 style={style}
                 size="small"
                 badge={true}
+                title={chip.iconTitle}
             />
         );
     }