+++ src/components/picker/picker.tsx
@@ -310,8 +310,16 @@ export class Picker {
 
     private renderSpinner() {
         return (
-            <div style={{ 'padding-top': '6px', 'text-align': 'center' }}>
-                <limel-spinner class={'dropdown--spinner'} />
+            <div
+                style={{
+                    width: '100%',
+                    display: 'flex',
+                    'align-items': 'center',
+                    'justify-content': 'center',
+                    padding: '1rem 0',
+                }}
+            >
+                <limel-spinner limeBranded={false} />
             </div>
         );
     }